refactor backoff
