+++ src/lib/openai.ts
@@ -1,6 +1,6 @@
 import { Configuration, OpenAIApi } from "openai";
 import "@logseq/libs";
-
+import { backOff } from "exponential-backoff";
 interface OpenAIOptions {
   apiKey: string;
   completionEngine?: string;
@@ -10,10 +10,28 @@ interface OpenAIOptions {
 
 const OpenAIDefaults = (apiKey: string): OpenAIOptions => ({
   apiKey,
-  completionEngine: 'text-davinci-002',
+  completionEngine: "text-davinci-002",
   temperature: 1.0,
   maxTokens: 1000,
 });
+
+const retryOptions = {
+  numOfAttempts: 3,
+  retry: (err: any) => {
+    if (!err.response || !err.response.data || !err.response.data.error) {
+      return false;
+    }
+    if (err.response.status === 429) {
+      const errorType = err.response.data.error.type;
+      if (errorType === "insufficient_quota") {
+        return false;
+      }
+      console.warn("Rate limit exceeded. Retrying...");
+      return true;
+    }
+    return false;
+  },
+};
 
 export async function openAI(
   input: string,
@@ -28,14 +46,18 @@ export async function openAI(
 
   const openai = new OpenAIApi(configuration);
 
-  const response = await openai.createCompletion(engine, {
-    prompt: input,
-    temperature: options.temperature,
-    max_tokens: options.maxTokens,
-    top_p: 1,
-    frequency_penalty: 0,
-    presence_penalty: 0,
-  });
+  const response = await backOff(
+    () =>
+      openai.createCompletion(engine, {
+        prompt: input,
+        temperature: options.temperature,
+        max_tokens: options.maxTokens,
+        top_p: 1,
+        frequency_penalty: 0,
+        presence_penalty: 0,
+      }),
+    retryOptions
+  );
 
   const choices = response.data.choices;
   if (choices && choices[0] && choices[0].text) {
